Created React components for form. Next to wire it up to Netify
Made it prettier! Also got Required prop working visually at least. [Might need to revisit Section since with Inner it doesn't work right when it's within the Article as it's set to 80%. Maybe it's fine though and I just need a PageSection component]

Added a section to the reset with additional preferences mainly so I can set sizing to border-box

Rebase squash number 2 with all the working Form components and styling